# swagger.doc.json
{
  "openapi": "3.0.0",
  "paths": {
    "/": {
      "get": {
        "operationId": "AppController_getHello",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/signup/{userType}": {
      "post": {
        "operationId": "AuthController_signup",
        "summary": "Sign up as BUYER or get a product key from '/api/auth/key' and signup as REALTOR to create a home and so on.",
        "parameters": [
          {
            "name": "userType",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string",
              "enum": ["BUYER", "REALTOR", "ADMIN"]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignupDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User successfully registered."
          },
          "401": {
            "description": "If Realtor user doesn't provide Product Key."
          },
          "409": {
            "description": "If user already exists."
          }
        },
        "tags": ["auth"]
      }
    },
    "/api/auth/signin": {
      "post": {
        "operationId": "AuthController_signin",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/signinDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Sign in user."
          },
          "400": {
            "description": "Password Incorrect."
          },
          "404": {
            "description": "User not found."
          }
        },
        "tags": ["auth"]
      }
    },
    "/api/auth/key": {
      "post": {
        "operationId": "AuthController_generateProductKey",
        "summary": "In production only current ADMIN can run this endpoint for generate product key and validate user as REALTOR or ADMIN",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/generateProductKeyDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User Validated successfully."
          }
        },
        "tags": ["auth"]
      }
    },
    "/api/auth/me": {
      "get": {
        "operationId": "AuthController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": "User Info",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserTokenInfo"
                }
              }
            }
          },
          "400": {
            "description": "Please sign in first!"
          }
        },
        "tags": ["auth"],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/api/homes": {
      "get": {
        "operationId": "HomeController_getHomes",
        "summary": "Get all homes and filter by query params.",
        "parameters": [
          {
            "name": "city",
            "required": false,
            "in": "query",
            "example": "Los Anglos",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "minPrice",
            "required": false,
            "in": "query",
            "example": "320000",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "maxPrice",
            "required": false,
            "in": "query",
            "example": "420000",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "propertyType",
            "required": false,
            "in": "query",
            "schema": {
              "enum": ["RESIDENTIAL", "CONDO"],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get all homes."
          },
          "404": {
            "description": "Home not found."
          }
        },
        "tags": ["homes"]
      },
      "post": {
        "operationId": "HomeController_createHome",
        "summary": "Realtor users can create home.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateHomeResponseDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Home craeted successfully."
          },
          "400": {
            "description": "For invalid inputs."
          },
          "401": {
            "description": "Please signup as realtor first."
          },
          "403": {
            "description": "You must be realtor to create home."
          }
        },
        "tags": ["homes"],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/api/homes/{id}": {
      "get": {
        "operationId": "HomeController_getOneHome",
        "summary": "Get home by id.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success."
          },
          "404": {
            "description": "Home not found."
          }
        },
        "tags": ["homes"]
      },
      "put": {
        "operationId": "HomeController_updateHome",
        "summary": "Realtor can update own home.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateHomeResponseDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Home updated successfully."
          },
          "400": {
            "description": "For invalid inputs."
          },
          "401": {
            "description": "Please signup as realtor first."
          },
          "403": {
            "description": "Only realtor users can update home. | Realtor id and home creator id didn't match."
          }
        },
        "tags": ["homes"],
        "security": [
          {
            "access-token": []
          }
        ]
      },
      "delete": {
        "operationId": "HomeController_deleteHome",
        "summary": "Realtor can delete own home.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Home deleted successfully."
          },
          "401": {
            "description": "Please signup as realtor first."
          },
          "403": {
            "description": "Only realtor users can delete home. | Realtor id and home creator id didn't match."
          }
        },
        "tags": ["homes"],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/api/homes/{id}/inquire": {
      "post": {
        "operationId": "HomeController_inquire",
        "summary": "Buyer can inquire for a home by messaging.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Id of the home.",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InquireDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Inqure home successfully."
          },
          "400": {
            "description": "For invalid inputs."
          },
          "401": {
            "description": "Please signup first."
          }
        },
        "tags": ["homes"],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/api/homes/{id}/messages": {
      "get": {
        "operationId": "HomeController_getHomeMessages",
        "summary": "Realtor can see all messages for specific home.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get all home messages."
          },
          "400": {
            "description": "For invalid inputs."
          },
          "401": {
            "description": "Please signup as realtor first."
          },
          "403": {
            "description": "Only realtor users can delete home. | Realtor id and home creator id didn't match."
          }
        },
        "tags": ["homes"],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    }
  },
  "info": {
    "title": "Realtor App",
    "description": "Realtor users can get the product key to post houses and Buyer users can contact to specific realtor of house.",
    "version": "1.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "Authorization",
        "in": "headers"
      }
    },
    "schemas": {
      "SignupDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Name of user",
            "example": "Jhon"
          },
          "phone": {
            "type": "string",
            "description": "Phone number",
            "example": "00989123456789"
          },
          "email": {
            "type": "string",
            "description": "Email",
            "example": "JhonDoe@example.com"
          },
          "password": {
            "type": "string",
            "minLength": 5,
            "description": "Password of user",
            "example": "password@123"
          },
          "productKey": {
            "type": "string",
            "description": "Some hashed token based on user info for Authenticate REALTORs and ADMINs.",
            "example": "SomeEncryptTokenBasedOnRealtorInfo"
          }
        },
        "required": ["name", "phone", "email", "password"]
      },
      "signinDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Email",
            "example": "JhonDoe@example.com"
          },
          "password": {
            "type": "string",
            "description": "Password of user",
            "example": "password@123"
          }
        },
        "required": ["email", "password"]
      },
      "generateProductKeyDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Email",
            "example": "JhonDoe@example.com"
          },
          "userType": {
            "enum": ["REALTOR", "ADMIN"],
            "type": "string"
          }
        },
        "required": ["email", "userType"]
      },
      "UserTokenInfo": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "id": {
            "type": "number"
          },
          "userType": {
            "type": "string"
          },
          "iat": {
            "type": "number"
          },
          "exp": {
            "type": "number"
          }
        },
        "required": ["name", "id", "userType", "iat", "exp"]
      },
      "CreateHomeResponseDto": {
        "type": "object",
        "properties": {
          "address": {
            "type": "string",
            "description": "Address of home",
            "example": "317, 2211 29 Street SW"
          },
          "numberOfBedrooms": {
            "type": "integer",
            "description": "Number of bedrooms",
            "example": 2
          },
          "numberOfBathrooms": {
            "type": "integer",
            "description": "Number of bathrooms",
            "example": 3
          },
          "city": {
            "type": "string",
            "example": "Los Anglos"
          },
          "price": {
            "type": "integer",
            "description": "Price of home",
            "example": 360000
          },
          "landSize": {
            "type": "number",
            "description": "Land size of home",
            "example": 6422
          },
          "propertyType": {
            "type": "string",
            "description": "Type of home",
            "enum": ["RESIDENTIAL", "CONDO"]
          },
          "images": {
            "description": "Array of images URLs",
            "example": [
              {
                "url": "https://cdn.realtor.ca/listing/TS637990644654170000/reb3/highres/2/914772_10.jpg"
              },
              {
                "url": "https://cdn.realtor.ca/listing/TS637990644623500000/reb3/highres/2/914772_5.jpg"
              },
              {
                "url": "https://cdn.realtor.ca/listing/TS637990644655930000/reb3/highres/2/914772_3.jpg"
              }
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "address",
          "numberOfBedrooms",
          "numberOfBathrooms",
          "city",
          "price",
          "landSize",
          "propertyType",
          "images"
        ]
      },
      "UpdateHomeResponseDto": {
        "type": "object",
        "properties": {
          "address": {
            "type": "string",
            "description": "New address of home",
            "example": "317, 2211 29 Street SW"
          },
          "numberOfBedrooms": {
            "type": "integer",
            "description": "Number of bedrooms",
            "example": 4
          },
          "numberOfBathrooms": {
            "type": "integer",
            "description": "Number of bathrooms",
            "example": 5
          },
          "city": {
            "type": "string",
            "example": "Babolsar"
          },
          "price": {
            "type": "number",
            "description": "New Price",
            "example": 1850000
          },
          "landSize": {
            "type": "number",
            "description": "Land size of home",
            "example": 16422
          },
          "propertyType": {
            "type": "string",
            "description": "Update type of home",
            "enum": ["RESIDENTIAL", "CONDO"]
          },
          "images": {
            "description": "Array of images URLs",
            "example": [
              "https://cdn.realtor.ca/listing/TS637990590487500000/reb5/highres/6/17979956_11.jpg",
              "https://cdn.realtor.ca/listing/TS637990590487470000/reb5/highres/6/17979956_10.jpg",
              "https://cdn.realtor.ca/listing/TS637990590488200000/reb5/highres/6/17979956_9.jpg"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "address",
          "numberOfBedrooms",
          "numberOfBathrooms",
          "city",
          "price",
          "landSize",
          "propertyType",
          "images"
        ]
      },
      "InquireDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Inquire message for home.",
            "example": "I like this home. How I can see it ?"
          }
        },
        "required": ["message"]
      }
    }
  }
}